Reflect "Call for Heat" Status
By checking if the heatingPower is set to 0 and setting the CurrentHeatingCoolingState to OFF, HomeKit now correctly indicates if the Tado is calling for heat or not, so its temperature icon goes green if the heating is on and it's not currently needing to fire the boiler, and orange if it's calling for heat and firing the boiler. It still displays the fully "Off" state if the heating is disabled.